Split transforms into multiple files
