+++ tests/test_heartbeat.py
@@ -3,7 +3,7 @@ import json
 import logging
 
 from aw_core.models import Event
-from aw_transform import transforms
+from aw_transform import heartbeat_merge, heartbeat_reduce
 
 import unittest
 
@@ -14,7 +14,7 @@ def test_heartbeat_merge():
     td_1s = timedelta(seconds=1)
 
     last_event, heartbeat = Event(timestamp=now), Event(timestamp=now + td_1s)
-    merged = transforms.heartbeat_merge(last_event, heartbeat, pulsetime=2)
+    merged = heartbeat_merge(last_event, heartbeat, pulsetime=2)
     assert merged is not None
 
 
@@ -25,12 +25,12 @@ def test_heartbeat_merge_fail():
 
     # timestamp of heartbeat more than pulsetime away
     last_event, heartbeat = Event(timestamp=now, data={"label": "test"}), Event(timestamp=now + 3*td_1s, data={"label": "test"})
-    merged = transforms.heartbeat_merge(last_event, heartbeat, pulsetime=2)
+    merged = heartbeat_merge(last_event, heartbeat, pulsetime=2)
     assert merged is None
 
     # labels not identical
     last_event, heartbeat = Event(timestamp=now, data={"label": "test"}), Event(timestamp=now + td_1s, data={"label": "test2"})
-    merged = transforms.heartbeat_merge(last_event, heartbeat, pulsetime=2)
+    merged = heartbeat_merge(last_event, heartbeat, pulsetime=2)
     assert merged is None
 
 
@@ -40,7 +40,7 @@ def test_heartbeat_reduce():
     td_1s = timedelta(seconds=1)
 
     events = [Event(timestamp=now, data={"label": "test"}), Event(timestamp=now + td_1s, data={"label": "test"})]
-    reduced_events = transforms.heartbeat_reduce(events, pulsetime=2)
+    reduced_events = heartbeat_reduce(events, pulsetime=2)
     assert len(reduced_events) == 1
 
 
@@ -50,5 +50,5 @@ def test_heartbeat_reduce_fail():
     td_1s = timedelta(seconds=1)
 
     events = [Event(timestamp=now, data={"label": "test"}), Event(timestamp=now + 3*td_1s, data={"label": "test"})]
-    reduced_events = transforms.heartbeat_reduce(events, pulsetime=2)
+    reduced_events = heartbeat_reduce(events, pulsetime=2)
     assert len(reduced_events) == 2
